requirements.txt
# DataForge.io Python Dependencies

# Core Framework
pydantic>=2.0
pyyaml==6.0
python-dotenv==1.0.0

# Apache Spark & Delta Lake
pyspark==3.5.0
delta-spark==3.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Streaming & Messaging
kafka-python==2.0.2

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
boto3==1.28.0  # For AWS S3
azure-storage-blob==12.18.0  # For Azure Blob Storage
minio==7.1.13  # For MinIO

# Web & API
flask==3.0.0
requests==2.31.0

# Observability & Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
pylint==2.17.5
black==23.9.1
isort==5.12.0
mypy==1.5.1

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-apache-kafka==5.4.0

# dbt
dbt-core==1.7.1
dbt-postgres==1.7.1
dbt-spark==1.7.1

# Utilities
click==8.1.7
typing-extensions==4.8.0
deepdiff==6.1.0

# Development
ipython==8.15.0
jupyter==1.0.0
